+++ NEWS.md
@@ -1,7 +1,7 @@
 # proxysiphon v0.0.1a4
 
 * Add `LgmRecord.chronology_depth_range()` to get the depth range of chronology determinants (Issue #11).
-
+* `LgmRecord.slice_datadepths()` now slices `data.age_ensemble` and `data.age_median` in addition to the general dataframe.
 
 # proxysiphon v0.0.1a3