+++ NEWS.md
@@ -1,3 +1,7 @@
+# proxysiphon v0.0.1a5
+
+*
+
 # proxysiphon v0.0.1a4
 
 * Add `LgmRecord.chronology_depth_range()` to get the depth range of chronology determinants (Issue #11).